+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## Features
 - Able to search time entries (#25)
 - Keyboard shortcut: Refresh Data (#24)
+- Current running shows on home page (#26)
+- Loading Screen (#32)
 
 ## Bugs
 - Graceful Shutdown on Error (#27)
 - Move Description to top of home (#28)
+- Cleanup (#29)